+++ packages/pangea-sdk/pangea/asyncio/services/authz.py
@@ -3,13 +3,15 @@
 
 from __future__ import annotations
 
+from collections.abc import Mapping, Sequence
 from typing import Any
 
 from pangea.asyncio.services.base import ServiceBaseAsync
 from pangea.config import PangeaConfig
 from pangea.response import PangeaResponse
 from pangea.services.authz import (
-    CheckRequest,
+    BulkCheckRequestItem,
+    BulkCheckResult,
     CheckResult,
     ItemOrder,
     ListResourcesRequest,
@@ -19,7 +21,6 @@ from pangea.services.authz import (
     Resource,
     Subject,
     Tuple,
-    TupleCreateRequest,
     TupleCreateResult,
     TupleDeleteRequest,
     TupleDeleteResult,
@@ -73,7 +74,7 @@ class AuthZAsync(ServiceBaseAsync):
 
         super().__init__(token, config, logger_name, config_id=config_id)
 
-    async def tuple_create(self, tuples: list[Tuple]) -> PangeaResponse[TupleCreateResult]:
+    async def tuple_create(self, tuples: Sequence[Tuple]) -> PangeaResponse[TupleCreateResult]:
         """Create tuples.
 
         Create tuples in the AuthZ Service. The request will fail if there is no schema
@@ -102,10 +103,7 @@ class AuthZAsync(ServiceBaseAsync):
             )
         """
 
-        input_data = TupleCreateRequest(tuples=tuples)
-        return await self.request.post(
-            "v1/tuple/create", TupleCreateResult, data=input_data.model_dump(exclude_none=True)
-        )
+        return await self.request.post("v1/tuple/create", TupleCreateResult, data={"tuples": tuples})
 
     async def tuple_list(
         self,
@@ -190,8 +188,8 @@ class AuthZAsync(ServiceBaseAsync):
         Check if a subject has permission to perform an action on the resource.
 
         Args:
-            resource (Resource): The resource to check.
-            action (str): The action to check.
+            resource: The resource to check.
+            action: The action to check.
             subject: The subject to check.
             debug: Setting this value to True will provide a detailed analysis of the check.
             attributes: Additional attributes for the check.
@@ -213,8 +211,44 @@ class AuthZAsync(ServiceBaseAsync):
             )
         """
 
-        input_data = CheckRequest(resource=resource, action=action, subject=subject, debug=debug, attributes=attributes)
-        return await self.request.post("v1/check", CheckResult, data=input_data.model_dump(exclude_none=True))
+        return await self.request.post(
+            "v1/check",
+            CheckResult,
+            data={"resource": resource, "action": action, "subject": subject, "debug": debug, "attributes": attributes},
+        )
+
+    async def bulk_check(
+        self,
+        checks: Sequence[BulkCheckRequestItem],
+        *,
+        debug: bool | None = None,
+        attributes: Mapping[str, Any] | None = None,
+    ) -> PangeaResponse[BulkCheckResult]:
+        """Perform a bulk check request
+
+        Perform multiple checks in a single request to see if a subjects have
+        permission to do actions on the resources.
+
+        Args:
+            checks: Check requests to perform.
+            debug: In the event of an allowed check, return a path that granted access.
+            attributes: A JSON object of attribute data.
+
+        Examples:
+            await authz.bulk_check(
+                checks=[
+                    BulkCheckRequestItem(
+                        resource=Resource(type="file", id="file_1"),
+                        action="read",
+                        subject=Subject(type="user", id="user_1", action="read"),
+                    )
+                ]
+            )
+        """
+
+        return await self.request.post(
+            "v1/check/bulk", BulkCheckResult, data={"checks": checks, "debug": debug, "attributes": attributes}
+        )
 
     async def list_resources(
         self, type: str, action: str, subject: Subject, attributes: dict[str, Any] | None = None